pyproject.toml
[tool.poetry]
name = "django-fahrenheit"
version = "0.1.0"
description = ""
authors = [
    "Mikhail Porokhovnichenko <marazmiki@gmail.com>"
]
# Look, mom: https://pypi.org/pypi?%3Aaction=list_classifiers
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 2.1",
    "Framework :: Django :: 2.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.7"
geoip2 = "^2.9"
django-generic-helpers = "^1.0"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
pytest-django = "^3.4"
pytest-env = "^0.6.2"
pytest-isort = "^0.3.1"
pytest-pythonpath = "^0.7.3"
pytest-flake8 = "^1.0"
pytest-cov = "^2.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
